27/10/2020

+++ lib/pet/models/pets.dart
@@ -4,14 +4,14 @@ class Pets {
   Pets.fromDocument(DocumentSnapshot document) {
     id = document.documentID;
     name = document['name'] as String;
-    idade = document['idade'] as double;
+    idade = document['idade'] as int;
     type = document['type'] as String;
     images = List<String>.from(document.data['images'] as List<dynamic>);
   }
 
   String id;
   String name;
-  double idade;
+  int idade;
   String type;
   List<String> images;
 }
